fix(release): align embedded schema version
