--- .github/workflows/repair-pr308.yml
@@ -1,23 +1,0 @@
-name: Repair PR 308 release metadata
-
-on:
-  push:
-    branches:
-      - agent/zoo-code-portable-integration
-
-permissions:
-  contents: write
-
-jobs:
-  repair:
-    if: github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v7
-        with:
-          fetch-depth: 0
-          ref: agent/zoo-code-portable-integration
-      - uses: actions/setup-python@v6
-        with:
-          python-version: '3.11'
-      - name
